refactor: refactored to get jest in working state
