+++ app.js
@@ -1,7 +1,10 @@
-import app from './server.js';
+import express from 'express';
+import apiRouter from './routes/api.js';
 
-const port = 3000;
+const app = express();
+app.use(express.json());
+app.use(express.urlencoded({ extended: false }));
 
-app.listen(port, () => {
-  console.log(`Server running on http://localhost:${port}`);
-});
+app.use('/api', apiRouter);
+
+export default app;
