+++ server.js
@@ -1,10 +1,7 @@
-import express from 'express';
-import router from './routes/api.js';
-const app = express();
+import app from './app.js';
 
-app.use(express.json());
-app.use(express.urlencoded({ extended: false }));
+const port = process.env.PORT | 3000;
 
-app.use('/api', router);
-
-export default app;
+app.listen(port, () =>
+  console.log(`Server running on port ${port}`)
+);
